ci: add a workflow to open component update PRs
Git for Windows already has a workflow to open tickets whenever
newer component versions are available (e.g. a newer OpenSSL or cURL
that needs to be packged and bundled with Git for Windows).

It makes only sense to automate even more of that process by adding a
workflow that can be triggered manually, specifying a ticket URL as
input that points to the ticket opened by above-mentioned workflow.

This will hopefully make it super easy eventually to build new packages
for the Git for Windows SDK whenever a new package version is available.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>